requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==5.0.1
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.0.0
elasticsearch-dsl==8.12.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.6.0
